WIP: refactor utils.
