+++ phy/utils/_misc.py
@@ -10,59 +10,12 @@
 import sys
 from inspect import getargspec
 
-import numpy as np
-
-from ..ext.six import string_types, integer_types
 from ..ext.six.moves import builtins
 
 
 #------------------------------------------------------------------------------
 # Various Python utility functions
 #------------------------------------------------------------------------------
-
-class Bunch(dict):
-    """A dict with additional dot syntax."""
-    def __init__(self, *args, **kwargs):
-        super(Bunch, self).__init__(*args, **kwargs)
-        self.__dict__ = self
-
-    def copy(self):
-        return Bunch(super(Bunch, self).copy())
-
-
-def _as_dict(x):
-    """Convert a list of tuples to a dict."""
-    if isinstance(x, list):
-        return dict(x)
-    else:
-        return x
-
-
-def _concatenate_dicts(*dicts):
-    """Concatenate dictionaries."""
-    out = {}
-    for dic in dicts:
-        out.update(dic)
-    return out
-
-
-def _is_list(obj):
-    return isinstance(obj, list)
-
-
-def _is_integer(x):
-    return isinstance(x, integer_types + (np.generic,))
-
-
-def _as_list(obj):
-    """Ensure an object is a list."""
-    if isinstance(obj, string_types):
-        return [obj]
-    elif not hasattr(obj, '__len__'):
-        return [obj]
-    else:
-        return obj
-
 
 def _fun_arg_count(f):
     """Return the number of arguments of a function.
